requirements.txt
loguru==0.7.3
fastapi==0.95.0
psycopg2-binary==2.9.10
python-dotenv==0.21.0
peewee==3.15.0
uvicorn==0.34.0
ruff==0.9.4
cerberus==1.3.7